requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
black==24.3.0
boto3==1.34.69
botocore==1.34.162
cbor2==5.8.0
celery==5.3.6
certifi==2026.1.4
cffi==2.0.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.13.4
crispy-bootstrap4==2024.1
cron_descriptor==2.0.6
cryptography==46.0.5
daphne==4.1.2
defusedxml==0.7.1
Django==4.2.11
django-allauth==0.61.1
django-cacheops==7.0.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-constance==3.1.0
django-cors-headers==4.3.1
django-countries==7.5.1
django-crispy-forms==2.3
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-haystack==3.3.0
django-model-utils==5.0.0
django-phonenumber-field==7.3.0
django-picklefield==3.4.0
django-redis==6.0.0
django-sslserver-v2==1.0
django-storages==1.14.2
django-timezone-field==7.2.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
flake8==7.0.0
funcy==2.0
gunicorn==21.2.0
h11==0.16.0
hiredis==3.3.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
jmespath==1.1.0
jsonfield==3.2.0
kombu==5.6.2
mccabe==0.7.0
msgpack==1.1.2
mypy_extensions==1.1.0
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==26.0
pathspec==1.0.4
phonenumbers==8.13.27
pillow==10.2.0
platformdirs==4.9.2
pluggy==1.6.0
prompt_toolkit==3.0.52
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==3.0
pyflakes==3.2.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
pytest==8.1.1
pytest-cov==4.1.0
pytest-django==4.8.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.2
redis==5.0.3
requests==2.32.5
requests-oauthlib==2.0.0
s3transfer==0.10.4
sentry-sdk==1.45.0
service-identity==24.2.0
setuptools==82.0.0
six==1.17.0
social-auth-app-django==5.4.0
social-auth-core==4.8.5
sqlparse==0.5.5
swapper==1.4.0
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.6.0
websockets==16.0
wheel==0.46.3
whitenoise==6.6.0
Whoosh==2.7.4
zope.interface==8.2
